+++ flash/video/classification/data.py
@@ -28,6 +28,7 @@ from flash.core.data.data_source import (
     PathsDataSource,
 )
 from flash.core.data.process import Preprocess
+from flash.core.integrations.labelstudio.data_source import LabelStudioVideoClassificationDataSource
 from flash.core.utilities.imports import _FIFTYONE_AVAILABLE, _KORNIA_AVAILABLE, _PYTORCHVIDEO_AVAILABLE, lazy_import
 
 SampleCollection = None
@@ -251,6 +252,13 @@ class VideoClassificationPreprocess(Preprocess):
                 ),
                 DefaultDataSources.FIFTYONE: VideoClassificationFiftyOneDataSource(
                     clip_sampler,
+                    video_sampler=video_sampler,
+                    decode_audio=decode_audio,
+                    decoder=decoder,
+                    **data_source_kwargs,
+                ),
+                DefaultDataSources.LABELSTUDIO: LabelStudioVideoClassificationDataSource(
+                    clip_sampler=clip_sampler,
                     video_sampler=video_sampler,
                     decode_audio=decode_audio,
                     decoder=decoder,